// src/ReduxTool/Slice/Consumer/types.ts
export interface initialstateType {
    isLoading:boolean,
    error:string|null
    orderModalOpen:boolean,
    projects: ProjectTy[] | null,
    roofAnalysis: roofAnalysisType,
    ondFiles:FileUplad | null
    orderstatusmodal:boolean
};
export interface roofAnalysisType {
    isWithRoof: boolean | null;
    isSaved: boolean;
    isSaveplantSummarydata:boolean;
    selected_project:ProjectTy | null;
    esriDraw:boolean;
    obsArea:string|null;
    showConfirmBtn: boolean;
    openNoRoof: boolean;
    formDetails: roofAnalysisFormType;
    regionName:string
    countryName:string
    selectedOption:string
    subsequentOption:string
    cardIconOpen:string
}
export interface roofAnalysisFormType {
    projectSetup: TProjectTypesNew;
    quickplantAnalysis: quickplantAnalysisType;
    projectSummary: ProjectSummaryType;
    error:any;
}
// ...
export type FileUplad = {  inverter: commomRes, module: commomRes }
// ...
export type FinRes = {
    montlyFinancialSaving: number,
    yearlyFinancialSaving: number,
    totalCarbonfootprint: number
    totalTrees: number
    netsavings: number
}